Add .travis.yml #1

+++ bin/tagging.sh
@@ -9,7 +9,7 @@ fi
 
 if [[ "master" != "$TRAVIS_BRANCH" ]]; then
   echo "Not on the 'master' branch."
-  #exit
+  exit
 fi
 
 if [[ "7" != "$TRAVIS_PHP_VERSION" ]]; then
@@ -17,16 +17,24 @@ if [[ "7" != "$TRAVIS_PHP_VERSION" ]]; then
 	exit
 fi
 
+curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+TAGNAME=$(php wp-cli.phar theme get earn-pocket-money --field=version)
+GIT_TAG_EXISTS=$(git tag -l '$TAGNAME')
+if [ -n "$GIT_TAG_EXISTS" ]; then
+  echo "tag $TAGNAME already exists."
+  exit
+fi
+
 git clone -b release --quiet git@github.com:earn-pocket-money/earn-pocket-money.git release
 cd release
 ls | xargs rm -rf
 ls -la
 rsync -auz --exclude="release" ../ .
-rm -rf .editorconfig .gitignore .travis.yml src gulpfile.js package.json yarn.lock node_modules
+rm -rf .editorconfig .gitignore .travis.yml .travis bin src gulpfile.js package.json yarn.lock node_modules
 ls -la
 
 git add -A
 git commit -m "[ci skip] release branch update from travis $TRAVIS_COMMIT"
 
-git tag $(wp theme get earn-pocket-money --field=version)
+git tag "$TAGNAME"
 git push origin release --tags
